Rename elrond::modules namespace to elrond::module
Issue #14
