+++ include/interface/Runtime.hpp
@@ -10,8 +10,8 @@
 
                 public:
 
-                    virtual elrond::modules::BaseGpioModule& getGpioService() const =0;
-                    virtual elrond::modules::BaseInputDriverModule& getInputService(const elrond::sizeT id = 0) const =0;
+                    virtual elrond::module::BaseGpioModule& getGpioService() const =0;
+                    virtual elrond::module::BaseInputDriverModule& getInputService(const elrond::sizeT id = 0) const =0;
                     virtual elrond::channel::BaseChannelManager& getChannelManager(const elrond::sizeT id = 0) const =0;
                     virtual const elrond::interface::DebugOut& dout() const =0;
