+++ include/module/BaseGpioModule.hpp
@@ -1,13 +1,12 @@
 #if !defined  _ELROND_BASE_GPIO_MODULE_HPP
     #define _ELROND_BASE_GPIO_MODULE_HPP
 
-    #include "types.hpp"
-    #include "modules/BaseModule.hpp"
+    #include "module/BaseModule.hpp"
 
     namespace elrond {
-        namespace modules {
+        namespace module {
 
-            class BaseGpioModule : public elrond::modules::BaseModule {
+            class BaseGpioModule : public elrond::module::BaseModule {
 
                 public:
 
@@ -16,8 +15,7 @@
                     #endif
 
                     virtual elrond::ModuleType getType() const override;
-
-                    virtual void attach(elrond::gpio::BaseGpioPin &pin)=0;
+                    virtual void attach(elrond::gpio::BaseGpioPin& pin)=0;
             };
         }
     }
