+++ include/module/BaseInputDriverModule.hpp
@@ -1,13 +1,12 @@
 #if !defined  _ELROND_BASE_INPUT_DRIVER_MODULE_HPP
     #define _ELROND_BASE_INPUT_DRIVER_MODULE_HPP
 
-    #include "types.hpp"
-    #include "modules/BaseModule.hpp"
+    #include "module/BaseModule.hpp"
 
     namespace elrond {
-        namespace modules {
+        namespace module {
 
-            class BaseInputDriverModule : public elrond::modules::BaseModule {
+            class BaseInputDriverModule : public elrond::module::BaseModule {
 
                 public:
 
@@ -15,7 +14,7 @@
                         virtual ~BaseInputDriverModule();
                     #endif
 
-                    virtual void addInputListener(const elrond::sizeT key, elrond::input::InputCallback *callback)=0;
+                    virtual void addInputListener(const elrond::sizeT key, elrond::input::InputCallback* callback)=0;
                     virtual elrond::ModuleType getType() const override;
             };
         }
