+++ include/module/ChannelToChannel.hpp
@@ -1,26 +1,27 @@
-#if !defined  _ELROND_LOOPBACK_MODULE_HPP
-    #define _ELROND_LOOPBACK_MODULE_HPP
+#if !defined  _ELROND_CHANNEL_TO_CHANNEL_MODULE_HPP
+    #define _ELROND_CHANNEL_TO_CHANNEL_MODULE_HPP
 
-    #include "types.hpp"
-    #include "modules/BaseTransportModule.hpp"
+    #include "module/BaseModule.hpp"
+    #include "channel/TxChannel.hpp"
+    #include "channel/RxChannel.hpp"
 
     namespace elrond {
-        namespace modules {
+        namespace module {
 
-            class Loopback : public modules::BaseTransportModule {
+            class ChannelToChannel : public module::BaseModule {
 
                 protected:
-                    elrond::channel::BaseChannelManager *cm = nullptr;
+                    bool inverted = false;
+                    elrond::channel::TxChannel txCh;
+                    elrond::channel::RxChannel rxCh;
 
                 public:
 
                     #ifdef ELROND_WITH_DESTRUCTORS
-                        virtual ~Loopback();
+                        virtual ~ChannelToChannel();
                     #endif
 
-                    void onInit(elrond::interface::ConfigMap& cfg) override;
-                    void send(elrond::byte data[], const elrond::sizeT length) override;
-                    void setChannelManager(elrond::channel::BaseChannelManager* cm) override;
+                    virtual void onInit(elrond::interface::ConfigMap& cfg) override;
 
                     #ifdef ELROND_WITH_MODULES_INFO
                         static const char *_getInternalName();
