+++ src/input/InputCallback.cpp
@@ -1,9 +1,9 @@
 #include "input/InputCallback.hpp"
 #include "interface/Runtime.hpp"
-#include "modules/BaseInputDriverModule.hpp"
+#include "module/BaseInputDriverModule.hpp"
 
 using elrond::input::InputCallback;
-using elrond::modules::BaseInputDriverModule;
+using elrond::module::BaseInputDriverModule;
 using elrond::input::OnInputHandleT;
 
 /*  ****************************************************************************
