+++ test/DigitalLed.test.cpp
@@ -8,7 +8,7 @@ using elrond::test::ChannelManagerTest;
 using elrond::test::ConfigMap;
 using elrond::test::DebugOut;
 
-using elrond::modules::DigitalLed;
+using elrond::module::DigitalLed;
 using elrond::gpio::BaseGpioPin;
 using elrond::gpio::DOutPin;
